provide the initial model-view-update structure
The Elm archiceture, otherwise known as TEA, is a model for handling interactive user interfaces. This concept was popularized in the frontend world by Redux, though Elm got there first. The idea is simple: every program has state, known as the model, a way of modifying the model, known as update, and finally a way to represent the model and allow for interactions known as the view. The model should be immutable - the update function should return a modified copy of the model. There are two many user-provided types: model, and message. The model type is state. The messages are triggered by interactions - for example, ClickOnSubmit.

This pattern is best applied to single page applications, where it makes sense to have a runtime that can render different content based on the model. In this initial commit, only the structure is provided. Later commits will introduce the flow that enables this pattern to be used.