add event handlers and patching
Now it's time to introduce event handlers to our AST. Each node can have events attached to it. An event is composed of a name of an event to listen for, and the callback that will take event data then produce a <message> that can be fed into the update loop. In order to modify the event handlers attached to the DOM elements, we store the event handlers with their IDs on the nodes. This allows us to remove the event handler. Patching is used in order to avoid fully rerendering the existing DOM, and instead only apply changes between the currently rendered version and the next.