add attributes to our Html representation
Attributes allow the DOM elements to change their settings. Attributes fall into two groups: attributes set via =, and properties set by modifying the properties of the DOM node in JavaScript. For now we just add support for attributes that have string values. Different tags support different attributes, but we'll leave that as an exercise for the reader.