introduce Html as an AST for rendering
To render content into the DOM, it's handy to have an Abstract Syntax Tree (AST) which represents the content to be rendered. In Elm, this is provided via the Html package. In React, JSX provides a user-familiar syntax while still turning into Reacty type stuff. In this example, we create our AST with 3 constructor functions that users can call - instead of using another templating language, make-your-own-tea focuses on everything being a function. Once we have our AST, we can use that the render and fill the DOM with actual elements.